Move `verifier.Verifier` to `block` to avoid import cycle
Block manipulation logic is needed in many unit tests. So is the
blockchain itself, which then again needs the block verification
implementation. To avoid import cycles, put block verifier into `block`
package.
